azure-functions/requirements.txt
azure-functions
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.5.0
yt-dlp>=2024.1.0
pyyaml>=6.0
requests>=2.31.0
PyGithub>=2.1.0
pydub>=0.25.1
